shree ganeshay namah , ssv no jay jay kar ho

+++ Routes/Executor/javaExecutor.js
@@ -1,64 +1,82 @@
 const compileRun = require('compile-run');
 const java=compileRun.java;
 const fs=require('fs');
-const path=require('path')
+const path=require('path');
+const {boundedProcessor} = require('../Processor/boundedProcessor');
 const {processor} = require('../Processor/processor')
 
-const javaExecutorLocally=async (dirpath,lang,code,input)=>{
+const javaExecutor=async (dirpath,lang,code,input,timelimit=5000,memorylimit=1048576)=>{
 
     // console.log("java : ",java);
 
-    let filename='Main.java';
-    let filepath=path.join(dirpath,filename);
-    fs.writeFileSync(filepath,code);
+    let sourceFileName='Main.java';
+    let sourceFilePath=path.join(dirpath,sourceFileName);
+    fs.writeFileSync(sourceFilePath,code);
 
     let inputFileName='input_java.txt';
     let inputFilePath=path.join(dirpath,inputFileName);
     fs.writeFileSync(inputFilePath,input);
 
-    const command='javac '+filepath;
+    const command='javac '+sourceFilePath;
     const compileResult=await processor(command);
     console.log("compileResult : ",compileResult);
     if(compileResult.stderr){
-        fs.unlinkSync(filepath);
-        fs.unlinkSync(inputFilePath);
+        try{
+            if (fs.existsSync(sourceFilePath)) fs.unlinkSync(sourceFilePath);
+        }catch(err){
+            console.log(err);
+        }
+        try{
+            if (fs.existsSync(inputFilePath)) fs.unlinkSync(inputFilePath);
+        }catch(err){
+            console.log(err);
+        }
         return {
             output:compileResult.stdout,
-            error:compileResult.stderr
+            error:compileResult.stderr,
+            executionTime:0.00
         }       
     }
-    const runcommand='java Main'+' < '+inputFilePath;
-    const result=await processor(runcommand);
+    const runcommand='java '+sourceFilePath+' < '+inputFilePath;
+    const result=await boundedProcessor(runcommand,timelimit,memorylimit);
 
     let output=result.stdout;
     let error=result.error;
-    
-    fs.unlinkSync(filepath);
-    fs.unlinkSync(inputFilePath);
+    let executionTime=result.executionTime;
+    try{
+        if (fs.existsSync(sourceFilePath)) fs.unlinkSync(sourceFilePath);
+    }catch(err){
+        console.log(err);
+    }
+    try{
+        if (fs.existsSync(inputFilePath)) fs.unlinkSync(inputFilePath);
+    }catch(err){
+        console.log(err);
+    }
+
     return {
         output:output,
-        error:error
+        error:error,
+        executionTime:executionTime
     }
 }
 
-const javaExecutor=async (dirpath,lang,code,input)=>{
+// const javaExecutor=async (dirpath,lang,code,input)=>{
 
-    console.log("java : ",java);
-    console.log("compile-run :",compileRun);
-    // let filename='Main.java';
-    // let filepath=path.join(dirpath,filename);
-    // fs.writeFileSync(filepath,code);
+//     console.log("java : ",java);
+//     console.log("compile-run :",compileRun);
+//     // let filename='Main.java';
+//     // let filepath=path.join(dirpath,filename);
+//     // fs.writeFileSync(filepath,code);
 
-    // let inputFileName='input_java.txt';
-    // let inputFilePath=path.join(dirpath,inputFileName);
-    // fs.writeFileSync(inputFilePath,input);
+//     // let inputFileName='input_java.txt';
+//     // let inputFilePath=path.join(dirpath,inputFileName);
+//     // fs.writeFileSync(inputFilePath,input);
 
-    return {
-        "java":java.hasOwnProperty("runFile")
-    }
-}
+//     return {
+//         "java":java.hasOwnProperty("runFile")
+//     }
+// }
 module.exports={
     javaExecutor:javaExecutor,
-    javaExecutorLocally:javaExecutorLocally
-
 }
